feat: Add sampling from p-star model using Metropolis dynamics
Sampling from p-star model (see Newman) is implemented. Rescaling of the parameters which makes Hamiltonian extensive in the number of nodes should be t_p --> t_p/(2*N_nodes)^p, which is different from what is currently implemented to match the rescaling from the draft of the paper. More efficient implementation based on adjacency matrix is probably possible.